TUC-1710 Only a rule can create and remove an DataServerStoppedAlarm. Removed call to do that based on a dataSourcePing().
Started working on a cluster sync point but the code is not being called anywhere.  This will be used to ensure that notifications are flushed from policy manager queues just after a state transition in order to avoid false triggering on stale notification data.  Not sure, yet, if this can happen, but having a means to check for it will help to track it down if there is an issue.


git-svn-id: svn://localhost/tmp/tungsten-replicator/trunk@1401 5c0c340a-b230-4711-bbfa-cee47acc84f3