Backported the following from the 2.0.1 branch.
TUC-1998 Recovery fails with NPE if replicator is offline. 
TUC-1997 Failover does not occur if replicator is offline.
Both of the above issues were due to the same problem: in the dataSourcePing() method, we were trying to get the db user, password, schema etc. from replicator properties.  A stopped or offline replicator does not provide these things properly.  I added some more properties to DataSource, which get populated from properties in the replicator when the DataSource gets created.  I then use the values of these properties to get the information we were getting from the replicator. This makes us replicator independent for failover etc.

git-svn-id: svn://localhost/tmp/tungsten-replicator/trunk@1942 5c0c340a-b230-4711-bbfa-cee47acc84f3