TUC-2206 This issue is due to a race condition that occurs when managers are stopped, on a given service, one at a time, and in the same order as they are started. In this case, what happens is that the first manager started becomes the coordinator, then the second, then the third. So when they are stopped in this order as well, what happens is that the next manager in the list becomes the coordinator and then does it's own validation of the remote composite service. When this happens, the coordinator will put the composite cluster in maintenance mode to facilitate the reconciliation of services. All is well here except the test is in the process of stopping all of the managers and so it ends up stopping a manager just after it has put the composite cluster into maintenance mode but before it is put back into automatic mode. When this occurs, the test fails because the remote cluster, whose job it is to shun the cluster on which we are stopping the managers, is in maintenance mode as well...
The fix for this involves a couple of things:

deferring the reconciliation of remote services for 5 seconds while increasing the remote service heartbeat interval to 10 seconds. In effect, this ensures that if the environment is in transition, that the reconciliation won't happen.
further, the rule that triggers to do the reconciliation won't trigger if there is a heartbeat gap alarm, which there will certainly be if other members of the cluster just dropped out.
Both of these things, together, resolve this issue. We do need a more robust means to maintain a distributed critical section with a means to commit/roll back operations during which the cluster membership changes.

Also cleaned up places where extraneous output was being generated in the manager log.

git-svn-id: svn://localhost/tmp/tungsten-replicator/branches/2.2.1-maintenance@2372 5c0c340a-b230-4711-bbfa-cee47acc84f3