TUC-2273 This issue turned out to be quite easy to fix. The 'epoch' in the HighWaterResource already represented the epoch_number for replication.  The problem was that the session HighWaterResource was not being updated with the latest epoch_number from the current master, but was only getting a copy of the epoch number from the datasource.  So the epoch number was never changing.
I also found but did not correct an issue with the call to TSRSession.updateHighWater()  This was getting called both at the time that we were about to use a connection and then afterwards if the connection was for a write. As far as I understand it, there's no reason to do this before using the write connection - only afterwards when it becomes significant because the highwater will represent the point at which the last write was done.  So in this way, the current code does three queries for every write i.e. the one on preparing to use the connection, another one for the users query, and the last after the write is done. This will REALLY slow things down.


git-svn-id: svn://localhost/tmp/tungsten-replicator/branches/2.2.1-maintenance@2494 5c0c340a-b230-4711-bbfa-cee47acc84f3