TUC-2273 I found the following: The root cause of this problem had to do with the way that datasources, in the connector, were being updated from datasources passed to the connector from the manager. During a switch operation, the DataSource is passed to the connector from the manager with the new role/state and the connector will update properties in a local copy of the DataSource with the properties it gets from the manager. The problem is that the 'highWater' property of DataSource in the manager is never initialized but when the DataSource is passed from the manager to the connector during a switch, etc. the uninitialized highWater from the manager clobbers the previously initialized highWater on the connector. This, in turn, led to the original reported issue.
This change eliminates the update of 'highWater' property of the connector DataSource.  This change also eliminates the select on the trep_commit_seqno.epoch_number column that was causing connector connections to be dropped if a 'lock tables' statement was used and smartScale was being used.

Finally, added a 'High water' column to the output of 'tungsten cluster status' since this will prove invaluable in debugging issues with SmartScale - like this one!


git-svn-id: svn://localhost/tmp/tungsten-replicator/trunk@2566 5c0c340a-b230-4711-bbfa-cee47acc84f3