Update Issue 910
Add --metadata-directory and --repl-metadata-directory options to store Continuent Tungsten metadata needed to ensure consistent operations between upgrades and configuration changes.

The --metadata-directory option will default to '{--home-directory}/metadata'.
The --repl-metadata-directory option will default to '{--metadata-directory}/applier'.

During the upgrade process, tpm will look for an existing 'tungsten-replicator/data' directory. If it exists, that directory will be moved to the path of '--repl-metadata-directory'. If the directory named by '--repl-metadata-directory' already exists, nothing will be moved and a warning will be displayed.


git-svn-id: svn://localhost/tmp/tungsten-replicator/trunk@2640 5c0c340a-b230-4711-bbfa-cee47acc84f3