CONT-26 - isolated relay site does not resume replication correctly CONT-29 - Manager not behaving correctly after the upgrade of the rules library CONT-54 Create composite datasource aborts if it cannot get the policy mode CONT-159 - Cluster remains in maintenance mode during concurrent recover/composite reconciliation
Resolved the issues above as well as a significant change for CONT-159 - a distributed critical section implementation. Due to this change, some Jenkins tests may fail due to executing a command that tries to acquire a critical section but times out.  This means that tests will need to look for the 'LOCK REQUEST TIMED OUT' text and retry the command if that happens. I will write more in a separate email.

git-svn-id: svn://localhost/tmp/tungsten-replicator/trunk@2716 5c0c340a-b230-4711-bbfa-cee47acc84f3